Add files via upload
Log 4: Although I determined the project was complete, further testing convinced me that this project was anything but done. However, after fixing the bugs that will be mentioned in this log, I feel comfortable with where the project is currently. As such, I will now focus on refactoring (a lot of it) and move on to the next project.